backend/requirements.txt
#app
fastapi==0.72.0
uvicorn==0.17.0
pydantic==1.9.0
email-validator==1.1.3
python-dotenv==0.19.2
python-multipart==0.0.5
sqladmin==0.10.3
pyjwt==2.6.0
sqlmodel==0.0.8

# db
SQLAlchemy==1.4.31
sqlalchemy-json==0.4.0
alembic==1.7.5
psycopg2-binary==2.9.3
motor==3.0.0